requirements.txt
numpy
Pillow
matplotlib
tqdm
av
pandas
pyarrow
torch
torchvision
wandb
